requirements.txt
torch
transformers
scikit-learn
numpy
umap-learn
plotly
click